更改：增加一些功能
