+++ main.py
@@ -4,18 +4,12 @@ script_path = os.path.split(os.path.realpath(__file__))[0]
 sys.path.append(script_path)
 from moligeek import *
 
-mainmode = input("请选择模式:\n[0]本机信息\n[1]web\n[2]network\n[3]LAN\n[4]密文处理\n")
+mainmode = input("请选择模式:\n[0]本机信息\n[1]web\n[2]network\n[3]LAN\n[4]密文解密\n[5]密文加密\n")
 if mainmode in ["0", "本机信息"]:
     hostinfo = network.hostinfo()
 elif mainmode in ["1", "web"]:
-    ssl = input("是否为https://\n[1]是\n[2]否\n")
-    if ssl in ["是", "1"]:
-        ssl = "https://"
-    elif ssl in ["否", "2"]:
-        ssl = "http://"
-    else:
-        print("未选择，默认为是")
-        ssl = "https://"
+    _ssl = input("是否为https://\n[1]是\n[2]否\n")
+    ssl = "https://" if _ssl in ["是", "1"] else "http://" if _ssl in ["否", "2"] else "https://"
     name = input("请输入链接:"+ssl)
     url = ssl+name
     mode = input("请选择模式:\n[1]下载源码\n[2]获取ip\n[3]提交表单\n[4]后台扫描\n[5]洪水攻击\n")
@@ -48,16 +42,21 @@ elif mainmode in ["3", "LAN"]:
         print("扫描中...")
         scan = LAN.scan(range)
 
-elif mainmode in ["4", "密文处理"]:
+elif mainmode in ["4", "密文解密"]:
     codetext = input("请输入文本:")
-    mode = input("请选择模式:\n[1]一键解密\n[2]栅栏解密\n[3]替代密码\n")
+    mode = input("请选择模式:\n[1]一键解密\n[2]栅栏解密\n[3]替代解密\n")
     if mode in ["1", "一键解密"]:
         encode = encode.todecode(codetext)
     elif mode in ["2", "栅栏解密"]:
         textlengh = int(input("每组字数:"))
         print(encode.fence.decrypt(codetext, textlengh))
     elif mode in ["3", "替代密码"]:
-        encode.Substitution.Decrypt(codetext)
+        encode.Tech.Substitution().Decrypt(codetext)
+elif mainmode in ["5", "密文加密"]:
+    codetext = input("请输入文本:")
+    mode = input("请选择模式:\n[1]替代加密\n[2]字典生成\n")
+    if mode in ["1", "替代加密"]:
+        encode.Tech.Substitution().Encrypt(codetext)
 
 if os.name == "nt":
     os.system("pause")
