+++ moligeek/core/encode/__init__.py
@@ -1,3 +1,3 @@
 from .decode import *
 from . import fence
-from . import Substitution
+from . import Tech
